+++ src/cu/warpsample.cuh
@@ -6,9 +6,9 @@
 
 namespace gplda {
 
-void compute_d_idx(cudaStream_t& stream, uint32_t* d_len, uint32_t* d_idx, size_t n_docs);
+void compute_d_idx(cudaStream_t& stream, uint32_t* d_len, uint32_t* d_idx, uint32_t n_docs);
 
-__global__ void warp_sample_topics(size_t size, size_t n_docs, uint32_t *z, uint32_t *w, uint32_t *d_len, uint32_t *d_idx, curandStatePhilox4_32_10_t* rng);
+__global__ void warp_sample_topics(uint32_t size, uint32_t n_docs, uint32_t *z, uint32_t *w, uint32_t *d_len, uint32_t *d_idx, curandStatePhilox4_32_10_t* rng);
 
 }
